Add NonFrontmostCollectionViewCellForItem
By examining this project, we can see how table/collection view behaves when you tell it to reload while it's not visible (e.g. under full screen modal). We see that, probably for performance reasons, it doesn't call cellForRow/cellForItem immediately. Instead, they will be called when collection view is visible again. Also it seems that it will be called not for all cell for which reload was called, but for those that are going to be visible on next appearance.